feat: バグ更新API作成
