+++ backend/app/forms/bug_form.rb
@@ -1,18 +1,14 @@
 class BugForm
   include ActiveModel::Model
   include ActiveModel::Validations
+  include BugFormValidations
 
-  attr_accessor :title, :error_message, :content, :expected_behavior, :solution, :cause, :etc, :is_solved, :status, :environments, :attempts, :references, :user
+  attr_accessor :title, :error_message, :content, :expected_behavior, :solution, :cause, :etc, :is_solved, :status, :environments, :attempts, :references,
+                :user, :bug
 
-  validates :title, presence: true, length: { maximum: 255 }
-  validates :content, presence: true
-  validate :solution_required_if_solved
-  validate :status_published_if_unsolved
-  validate :valid_references_urls
-  validate :valid_environments_name_and_version
-
-  def initialize(bug_params, user)
+  def initialize(bug_params, user, bug = nil)
     @user = user
+    @bug = bug
     @title = bug_params[:title]
     @error_message = bug_params[:error_message]
     @content = bug_params[:content]
@@ -31,17 +27,11 @@ class BugForm
     return false unless valid?
 
     ActiveRecord::Base.transaction do
-      bug = @user.bugs.create!(
-        title: @title, error_message: @error_message, content: @content,
-        expected_behavior: @expected_behavior, solution: @solution,
-        cause: @cause, etc: @etc, is_solved: @is_solved, status: @status
-      )
-
-      import_environments(bug)
-      import_attempts(bug)
-      import_references(bug)
-
-      bug
+      @bug ? update_bug : create_bug
+      update_environments(@bug)
+      update_attempts(@bug)
+      update_references(@bug)
+      @bug
     end
   rescue ActiveRecord::RecordInvalid
     false
@@ -49,19 +39,24 @@ class BugForm
 
   private
 
-    def solution_required_if_solved
-      if is_solved && solution.blank?
-        errors.add(:solution, "解決済にするには解決方法を入力する必要があります")
-      end
+    def update_bug
+      @bug.update!(
+        title: @title, error_message: @error_message, content: @content,
+        expected_behavior: @expected_behavior, solution: @solution,
+        cause: @cause, etc: @etc, is_solved: @is_solved, status: @status
+      )
     end
 
-    def status_published_if_unsolved
-      if !is_solved && status == "published"
-        errors.add(:status, "未解決の場合公開できません")
-      end
+    def create_bug
+      @bug = @user.bugs.create!(
+        title: @title, error_message: @error_message, content: @content,
+        expected_behavior: @expected_behavior, solution: @solution,
+        cause: @cause, etc: @etc, is_solved: @is_solved, status: @status
+      )
     end
 
-    def import_environments(bug)
+    def update_environments(bug)
+      bug.environments.delete_all
       environments_data = @environments.map do |env|
         { category: env[:category], name: env[:name], version: env[:version], bug_id: bug.id }
       end
@@ -69,7 +64,8 @@ class BugForm
       bug.environments.import(environments_data, on_duplicate_key_ignore: true) if environments_data.present?
     end
 
-    def import_attempts(bug)
+    def update_attempts(bug)
+      bug.attempts.delete_all
       attempts_data = @attempts.map do |attempt|
         { content: attempt[:content], bug_id: bug.id }
       end
@@ -77,36 +73,12 @@ class BugForm
       bug.attempts.import(attempts_data, on_duplicate_key_ignore: true) if attempts_data.present?
     end
 
-    def import_references(bug)
+    def update_references(bug)
+      bug.references.delete_all
       references_data = @references.map do |reference|
         { url: reference[:url], bug_id: bug.id }
       end
 
       bug.references.import(references_data, on_duplicate_key_ignore: true) if references_data.present?
-    end
-
-    def valid_references_urls
-      @references.each_with_index do |reference, _index|
-        url = reference[:url]
-        next if url.blank?
-
-        unless url.length <= 255
-          errors.add(:'references.url', "参考リンクは255文字以内で入力してください")
-        end
-        unless url =~ %r{\Ahttps?://[\S]+\z}
-          errors.add(:'references.url', "有効なURLを入力してください")
-        end
-      end
-    end
-
-    def valid_environments_name_and_version
-      @environments.each do |env|
-        if env[:name].present? && env[:name].length > 255
-          errors.add(:'environments.name', "環境の名前は255文字以内で入力してください")
-        end
-        if env[:version].present? && env[:version].length > 255
-          errors.add(:'environments.version', "環境のバージョンは255文字以内で入力してください")
-        end
-      end
     end
 end
