+++ backend/config/routes.rb
@@ -10,7 +10,7 @@ Rails.application.routes.draw do
         resource :confirmations, only: [:update]
       end
       resources :categories, only: [:index]
-      resources :bugs, only: [:index, :show, :create, :destroy]
+      resources :bugs, only: [:index, :show, :create, :update, :destroy]
     end
   end
